pyproject.toml
[project]
name = "sololc-vvault"
version = "0.1.2"
description = "vlt - Minimalist Secure Vault (Vvault CLI)"
readme = "README.md"
license = { text = "BSD-3-Clause" }
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
authors = [
    { name = "Josh Klein", email = "gigacoderx@gmail.com" }
]
requires-python = ">=3.11.15"
dependencies = [
    "cryptography>=46.0.7",
    "pillow>=12.2.0",
    "pyotp>=2.9.0",
    "pyperclip>=1.11.0",
    "pyyaml>=6.0.3",
    "pyzbar>=0.1.9",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "qrcode>=8.2",
]

[project.urls]
Homepage = "https://github.com/nuv0x/sololc-vvault"
Documentation = "https://github.com/nuv0x/sololc-vvault#readme"
Repository = "https://github.com/nuv0x/sololc-vvault"
Issues = "https://github.com/nuv0x/sololc-vvault/issues"
Sponsor = "https://ko-fi.com/sololc"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
vlt = "sololc_vvault:app"  # Key: Map the vlt command to your program


